Split expression aggregate checking
